+++ packages/components/src/synergy.ts
@@ -7,6 +7,8 @@
 export { default as SynAccordion } from './components/accordion/accordion.js';
 export { default as SynAlert } from './components/alert/alert.js';
 export { default as SynBadge } from './components/badge/badge.js';
+export { default as SynBreadcrumbItem } from './components/breadcrumb-item/breadcrumb-item.js';
+export { default as SynBreadcrumb } from './components/breadcrumb/breadcrumb.js';
 export { default as SynButtonGroup } from './components/button-group/button-group.js';
 export { default as SynButton } from './components/button/button.js';
 export { default as SynCard } from './components/card/card.js';